Created full CRUD: created  necessary services, changed all ids to long, improved tests according changes

+++ src/main/java/com/epam/note/dao/NoteRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoteRepository extends JpaRepository<NoteEntity, Integer> {
-    List<NoteEntity> findAllByNotebookId(int notebookId);
+public interface NoteRepository extends JpaRepository<NoteEntity, Long> {
+
+  List<NoteEntity> findAllByNotebookId(long notebookId);
+
 }
